backend/requirements.txt
Django==5.2.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
psycopg[binary]==3.2.10
python-dotenv==1.0.1
pytest==8.4.2
pytest-django==4.11.1
django-cors-headers==4.9.0
django-filter==24.3
google-auth==2.33.0
requests==2.32.3
gunicorn==23.0.0
whitenoise==6.8.2
dj-database-url==2.2.0
